flask_app/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.1.0
flatbuffers==25.2.10
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pyreadline3==3.5.4
sympy==1.13.3
Werkzeug==3.1.3
